Rename ClientKeyAuthenticationAttribute to SharedAccessKeyAuthenticationAttribute
