+++ src/Codit.Connectors.ApplicationInsights/Codit.Connectors.ApplicationInsights.csproj
@@ -141,7 +141,7 @@
     <Compile Include="ExceptionHandling\OwinExceptionHandler.cs" />
     <Compile Include="Extensions\IDictionaryExtensions.cs" />
     <Compile Include="Filters\AuthenticationFailureResult.cs" />
-    <Compile Include="Filters\ClientKeyAuthenticationAttribute.cs" />
+    <Compile Include="Filters\SharedAccessKeyAuthenticationAttribute.cs" />
     <Compile Include="Guard.cs" />
     <Compile Include="ExceptionHandling\ApplicationInsightsExceptionLogger.cs" />
     <Compile Include="OwinStartup.cs" />
